pyproject.toml
[tool.poetry]
name = "learn-fastapi"
version = "0.1.0"
description = ""
authors = ["https://github.com/alllexxx1"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.112.2"}
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
celery = "^5.4.0"
flower = "^2.0.1"
pillow = "^11.0.0"
sqladmin = "^0.20.1"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
